Move service worker back to src/
