+++ tsconfig.json
@@ -21,6 +21,6 @@
     "jsx": "react"
   },
   "include": [
-    "src", "public/serviceWorker.ts"
+    "src", "src/serviceWorker.ts"
   ]
 }
